:shell: add docker-run shared launcher, remove .sh from launchers

+++ launcher/docker-run.sh
@@ -1,8 +1,14 @@
 #!/bin/bash
 
-WINE_DOCKER="6.0.3"
-ROOT_DIR="/home/thomas/games/ftl"
-RUN_EXE="FTLGame.EXE"
+if [ ! -d "$ROOT_DIR" ] ; then
+  echo "ROOT_DIR '$ROOT_DIR' is missing!"
+  exit 1
+fi
+
+if [ ! -f "$ROOT_DIR/app/$RUN_EXE" ] ; then
+  echo "RUN_EXE '$RUN_EXE' is missing!"
+  exit 1
+fi
 
 docker run --rm -it \
   --name wine \
@@ -12,6 +18,7 @@ docker run --rm -it \
   --ipc="host" \
   -e USER_ID=$(id -u) \
   -e VIDEO_GID=$(  cat /etc/group | grep video  | cut -f3 -d":" ) \
+  -e WINETRICKS="${WINETRICKS:-isolate_home}" \
   -e HOME=/home/wine \
   -e DISPLAY=unix$DISPLAY \
   -v /tmp/.X11-unix:/tmp/.X11-unix \
@@ -22,6 +29,5 @@ docker run --rm -it \
   -v $ROOT_DIR/user:/home/wine/user \
   -v $ROOT_DIR/wineprefix:/home/wine/.wine \
   -w /home/wine/app \
-  0lfi/wine:$WINE_DOCKER \
+  0lfi/wine:${WINE_DOCKER:-latest} \
   bash -c "su wine -p -c \"wine $RUN_EXE"\"
-#  -e WINETRICKS="$WINETRICKS" \
